+++ app/db/models/courier_wallet.py
@@ -2,7 +2,7 @@
 Courier Wallet Model - Balance Tracking
 """
 from datetime import datetime
-from sqlalchemy import Column, Integer, Float, DateTime, ForeignKey
+from sqlalchemy import Column, Integer, BigInteger, Float, DateTime, ForeignKey
 from sqlalchemy.orm import relationship
 
 from app.db.database import Base
@@ -14,7 +14,7 @@ class CourierWallet(Base):
     __tablename__ = "courier_wallets"
 
     id = Column(Integer, primary_key=True, index=True)
-    courier_id = Column(Integer, ForeignKey("users.id"), unique=True, nullable=False)
+    courier_id = Column(BigInteger, ForeignKey("users.id"), unique=True, nullable=False)
 
     balance = Column(Float, default=0.0)
     credit_limit = Column(Float, default=-100.0)  # Minimum allowed balance
